backend/requirements.txt
flask
flask[async]
Flask-Cors
telethon
flask-jwt-extended
openai
requests
google-generativeai
Flask-SQLAlchemy>=2.5.0

pytest>=7.0.0
pytest-flask>=1.2.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0